% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{cdfPerPvalue}
\alias{cdfPerPvalue}
\title{CDF of p-values at selected thresholds}
\usage{
cdfPerPvalue(
  subFrame,
  pvalues = c(10^(-10:-2), 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.1)
)
}
\arguments{
\item{subFrame}{A vibration data frame subset.}

\item{pvalues}{Numeric vector of p-value thresholds.}
}
\value{
A data frame with columns \code{pvalue}, \code{cdf}, and \code{number}.
}
\description{
CDF of p-values at selected thresholds
}
\keyword{internal}
