% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metaanalysis.R
\name{clean_metaanalysis}
\alias{clean_metaanalysis}
\title{Clean meta-analysis output and get summary statistics.}
\usage{
clean_metaanalysis(metaanalysis, logger)
}
\arguments{
\item{metaanalysis}{Meta-analysis output.}

\item{logger}{Logger object (default = NULL).}
}
\description{
Export meta-analysis.
}
\keyword{meta-analysis}
