% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibration.R
\name{column_headers}
\alias{column_headers}
\title{Standardize column names for vibration output}
\usage{
column_headers(vibFrame, family)
}
\arguments{
\item{vibFrame}{A vibration results matrix.}

\item{family}{Model family.}
}
\value{
Character vector of standardized column names.
}
\description{
Standardize column names for vibration output
}
\keyword{internal}
