% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voe_pipeline.R
\name{full_voe_pipeline}
\alias{full_voe_pipeline}
\title{The VOE Pipeline}
\usage{
full_voe_pipeline(
  dependent_variables,
  independent_variables,
  primary_variable,
  fdr_method = "BY",
  fdr_cutoff = 0.05,
  max_vibration_num = 50000,
  proportion_cutoff = 0.1,
  meta_analysis = FALSE
)
}
\arguments{
\item{dependent_variables}{A tibble containing the information for your dependent variables (e.g. bacteria relative abundance, age). The first column should be the rownames (e.g. gene1, gene2, gene3), and the columns should correspond to different samples (e.g. individual1, individual2, etc).}

\item{independent_variables}{A tibble containing the independent variables you will want to vibrate over. Each column should correspond to a different variable (e.g. age), with the first column containing the sample names matching those in the column anmes of the dependent_variables tibble.}

\item{primary_variable}{The column name from the independent_variables tibble containing the key variable you want to associate with disease in your first round of modeling (prior to vibration). For example, if you are interested fundamentally identifying how well age can predict height, you would make this value a string referring to whatever column in said dataframe refers to "age."}

\item{fdr_method}{Your choice of method for adjusting p-values. Options are BY (default), BH, or bonferroni.}

\item{fdr_cutoff}{Cutoff for an FDR significant association (default = 0.05).}

\item{max_vibration_num}{Maximum number of vibrations (default=50000).}

\item{meta_analysis}{TRUE/FALSE -- indicates if computing meta-analysis across multiple datasets.}
}
\description{
This function will run the full pipeline
}
\examples{
voepipeline(metadata, abundance_data, mapping)
}
\keyword{pipeline}
