% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibration.R
\name{gatherFrames}
\alias{gatherFrames}
\title{Gather and harmonize all vibration results}
\usage{
gatherFrames(returnFrames)
}
\arguments{
\item{returnFrames}{List of results from \code{\link[=conductVibrationForK]{conductVibrationForK()}}.}
}
\value{
A list with \code{vibFrame}, \code{bicFrame}, \code{combinations}, \code{adjust},
\code{family}, and \code{base_formula}.
}
\description{
Gather and harmonize all vibration results
}
\keyword{internal}
