% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibration.R
\name{gatherVibration}
\alias{gatherVibration}
\title{Gather vibration results across k values}
\usage{
gatherVibration(returnFrames)
}
\arguments{
\item{returnFrames}{List of results from \code{\link[=conductVibrationForK]{conductVibrationForK()}}.}
}
\value{
A matrix of vibration results with a \code{k} column appended.
}
\description{
Gather vibration results across k values
}
\keyword{internal}
