% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibration.R
\name{gatherVibrationBIC}
\alias{gatherVibrationBIC}
\title{Gather BIC results across k values}
\usage{
gatherVibrationBIC(returnFrames)
}
\arguments{
\item{returnFrames}{List of results from \code{\link[=conductVibrationForK]{conductVibrationForK()}}.}
}
\value{
A matrix of BIC results with a \code{k} column appended.
}
\description{
Gather BIC results across k values
}
\keyword{internal}
