% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_linear_output.R
\name{get_adjuster_expanded_vibrations}
\alias{get_adjuster_expanded_vibrations}
\title{Expand vibration data}
\usage{
get_adjuster_expanded_vibrations(voe_df, adjusters, logger)
}
\arguments{
\item{voe_df}{Raw vibration output, the first entry in the output of compute_vibrations.}

\item{adjusters}{A dataframe, each column corresponding to the adjusters used in each dataset for vibrations. This is the second entry in the compute_vibrations output.}

\item{logger}{Logger object (default = NULL).}
}
\description{
Add column to vibration output for each adjuster (indicating presence or absence in a model).
}
\examples{
get_adjuster_expanded_vibrations(voe_df, adjusters,logger)
}
\keyword{analysis}
\keyword{voe}
