% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibration.R
\name{getBIC}
\alias{getBIC}
\title{Compute BIC for a model}
\usage{
getBIC(mod)
}
\arguments{
\item{mod}{A fitted model object.}
}
\value{
Numeric vector with effective degrees of freedom and BIC.
}
\description{
Compute BIC for a model
}
\keyword{internal}
