% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibration.R
\name{harmonizeFrame}
\alias{harmonizeFrame}
\title{Harmonize vibration frame column names and add derived columns}
\usage{
harmonizeFrame(vibFrame, family)
}
\arguments{
\item{vibFrame}{A vibration results matrix.}

\item{family}{Model family.}
}
\value{
A data frame with standardized column names.
}
\description{
Harmonize vibration frame column names and add derived columns
}
\keyword{internal}
