% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_logger.R
\name{initialize_logger}
\alias{initialize_logger}
\title{Logger}
\usage{
initialize_logger(fileName, saveLog = TRUE, logFilePath = NULL)
}
\arguments{
\item{fileName}{Name of logfile}

\item{saveLog}{TRUE/FALSE to save the logfile or not}

\item{logFilePath}{if saveLog=TRUE, path to directory where logfile will be saved.}
}
\description{
Initialize logging
}
\keyword{logging}
