% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibration.R
\name{int_to_binary_vector}
\alias{int_to_binary_vector}
\title{Convert integer to binary vector}
\usage{
int_to_binary_vector(x, n)
}
\arguments{
\item{x}{Integer to convert.}

\item{n}{Length of resulting binary vector.}
}
\value{
Integer vector of 0s and 1s.
}
\description{
Convert integer to binary vector
}
\keyword{internal}
