% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{meanEstimate}
\alias{meanEstimate}
\title{Median estimate and p-value}
\usage{
meanEstimate(subFrame)
}
\arguments{
\item{subFrame}{A vibration data frame subset.}
}
\value{
A one-row data frame with \code{estimate} and \code{pvalue}.
}
\description{
Median estimate and p-value
}
\keyword{internal}
