% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metaanalysis.R
\name{meta_analysis_command}
\alias{meta_analysis_command}
\title{Meta-analysis function}
\usage{
meta_analysis_command(a, b, c)
}
\arguments{
\item{a}{estimates}

\item{b}{standard errors}

\item{c}{study labels}
}
\description{
metagen function
}
\keyword{meta-analysis}
