% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{post_process}
\alias{post_process}
\title{Post-processing and summary statistics for VoE results}
\description{
Functions to compute summary statistics from vibration output, including
quantiles, relative hazard ratios (RHR), and relative p-values (RP).
}
