% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{quantilesEstimate}
\alias{quantilesEstimate}
\title{Quantiles of estimates}
\usage{
quantilesEstimate(subFrame, probs)
}
\arguments{
\item{subFrame}{A vibration data frame subset.}

\item{probs}{Probability quantiles.}
}
\value{
A data frame with columns \code{probs} and \code{estimate}.
}
\description{
Quantiles of estimates
}
