% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{quantilesHR}
\alias{quantilesHR}
\title{Quantiles of hazard ratios}
\usage{
quantilesHR(subFrame, probs = c())
}
\arguments{
\item{subFrame}{A vibration data frame subset.}

\item{probs}{Probability quantiles.}
}
\value{
A data frame with columns \code{probs} and \code{HR}.
}
\description{
Quantiles of hazard ratios
}
