% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{quantilesPvalue}
\alias{quantilesPvalue}
\title{Quantiles of p-values}
\usage{
quantilesPvalue(subFrame, probs = c(0.01, 0.25, 0.5, 0.75, 0.99))
}
\arguments{
\item{subFrame}{A vibration data frame subset.}

\item{probs}{Probability quantiles.}
}
\value{
A data frame with columns \code{probs} and \code{pvalue}.
}
\description{
Quantiles of p-values
}
