% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibration.R
\name{run_model}
\alias{run_model}
\title{Fit a model}
\usage{
run_model(form, data, family = "gaussian", ...)
}
\arguments{
\item{form}{A formula.}

\item{data}{A data frame.}

\item{family}{One of \code{"gaussian"}, \code{"cox"}, or \code{"binomial"}.}

\item{...}{Additional arguments passed to the model fitting function.}
}
\value{
A fitted model object.
}
\description{
Dispatches to \code{\link[=lm]{lm()}}, \code{\link[survival:coxph]{survival::coxph()}}, or \code{\link[=glm]{glm()}} based on family.
}
