% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{statPerKandFactor}
\alias{statPerKandFactor}
\title{Summary statistics per k and factor level}
\usage{
statPerKandFactor(vibFrame)
}
\arguments{
\item{vibFrame}{A vibration data frame.}
}
\value{
A data frame with columns \code{k}, \code{estimate}, \code{pvalue}, and
\code{factor_level}.
}
\description{
Summary statistics per k and factor level
}
