% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vibration.R
\name{vib2d_cox}
\alias{vib2d_cox}
\title{2D hexbin volcano plot for Cox VoE}
\usage{
vib2d_cox(vibObj, factor_num = 1, nbins = 20)
}
\arguments{
\item{vibObj}{A VoE result object (output of \code{\link[=conductVibration]{conductVibration()}} or
\code{\link[=conductVibrationSample]{conductVibrationSample()}}).}

\item{factor_num}{Factor level to plot (default 1).}

\item{nbins}{Number of hexbin bins (default 20).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
2D hexbin volcano plot for Cox VoE
}
\keyword{internal}
