% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vibration.R
\name{vibcontour_cox}
\alias{vibcontour_cox}
\title{Contour volcano plot for Cox VoE}
\usage{
vibcontour_cox(vibObj, factor_num = 1, alpha = 1)
}
\arguments{
\item{vibObj}{A VoE result object.}

\item{factor_num}{Factor level to plot (default 1).}

\item{alpha}{Point transparency (default 1).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Contour volcano plot for Cox VoE
}
\keyword{internal}
