% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibration.R
\name{vibration}
\alias{vibration}
\title{Vibration of Effects}
\description{
Core functions for computing the Vibration of Effects (VoE).
}
\references{
Patel CJ, Burford B, Ioannidis JPA. Assessment of vibration of effects due
to model specification can demonstrate the instability of observational
associations. \emph{Journal of Clinical Epidemiology}. 2015;68(9):1046-1058.
}
