% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_logger.R
\name{initialize_logger}
\alias{initialize_logger}
\title{Logger}
\usage{
initialize_logger(fileName, saveLog, logFilePath)
}
\description{
Initialize logging
#' @param fileName
#' @param saveLog 
#' @param logFilePath
}
\examples{
ind_var_analysis(metadata)
}
\keyword{logging}
